+++ test/function/arithmetic/divide.test.js
@@ -62,7 +62,7 @@ describe('divide', function() {
   });
 
   it('should divide each elements in a matrix by a number', function() {
-    assert.deepEqual(divide(math.range(2,2,7), 2), [1,2,3]);
+    assert.deepEqual(divide(math.range(2,8,2), 2), [1,2,3]);
     a  = math.matrix([[1,2],[3,4]]);
     assert.deepEqual(divide(a, 2), math.matrix([[0.5,1],[1.5,2]]));
     assert.deepEqual(divide(a.valueOf(), 2), [[0.5,1],[1.5,2]]);