+++ test/function/statistics/max.test.js
@@ -11,7 +11,7 @@ assert.equal(math.max('A', 'C', 'D', 'B'), 'D');
 assert.equal(math.max([1,3,5,2,-5]), 5);
 assert.equal(math.max(math.matrix([1,3,5,2,-5])), 5);
 assert.equal(math.max(math.range(1,6)), 5);
-assert.equal(math.max(math.range(5,-1,1)), 5);
+assert.equal(math.max(math.range(5,1,-1)), 5);
 assert.throws(function() {math.max()});
 assert.throws(function() {math.max([5,2], 3)});
 assert.throws(function() {math.max([])});